+++ src/lib.rs
@@ -23,5 +23,8 @@ pub mod geometry;
 pub(crate) mod model;
 pub mod scene;
 
+#[cfg(test)]
+mod test_support;
+
 pub use error::Error;
 pub use model::{edge::EdgeId, node::NodeId};